backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
PyJWT==2.8.0